requirements.txt
grpcio-tools==1.10.0
grpcio==1.10.0
requests==2.18.4
